Replace Snyk badge with Qodana in README.md
The Snyk badge used for showing security status in README.md has been replaced with a Qodana badge. This update aligns the README.md with the recent shift in our code quality monitoring from Snyk to Qodana. It maintains an accurate and up-to-date representation of our project's code quality measures.