Add SMS inbound action implementations and tests
This commit adds new classes for handling inbound SMS actions, including the implementation for the SmsInboundAction interface and a factory for creating instances of it. This also introduces testing of provider factory initialization, ensuring it correctly gets instances of the inbound action factories. A new dependency version has been defined in the pom.xml file as well.